pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "localllmrag"
version = "0.0.0"
description = "Local LLM RAG appliction for codebases"
authors = [{name = "Ryan McLean", email = "pvtryan100@gmail.com"}]
dependencies = [
    "aiohappyeyeballs==2.4.6",
    "aiohttp==3.11.12",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "asgiref==3.8.1",
    "astor==0.8.1",
    "attrs==25.1.0",
    "backoff==2.2.1",
    "bcrypt==4.2.1",
    "cachetools==5.5.1",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "chroma-hnswlib==0.7.6",
    "chromadb==0.6.3",
    "clang==17.0.6",
    "click==8.1.8",
    "coloredlogs==15.0.1",
    "dataclasses-json==0.6.7",
    "Deprecated==1.2.18",
    "durationpy==0.9",
    "EditorConfig==0.17.0",
    "fastapi==0.115.8",
    "filelock==3.17.0",
    "flatbuffers==25.1.24",
    "frozenlist==1.5.0",
    "fsspec==2025.2.0",
    "goastpy==0.1.10",
    "google-auth==2.38.0",
    "googleapis-common-protos==1.66.0",
    "grpcio==1.70.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.28.1",
    "humanfriendly==10.0",
    "idna==3.10",
    "importlib_metadata==8.5.0",
    "importlib_resources==6.5.2",
    "iniconfig==2.0.0",
    "javalang==0.13.0",
    "Jinja2==3.1.5",
    "joblib==1.4.2",
    "jsbeautifier==1.15.3",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.23.0",
    "jsonschema-specifications==2024.10.1",
    "kubernetes==32.0.0",
    "langchain==0.3.18",
    "langchain-chroma==0.2.1",
    "langchain-community==0.3.17",
    "langchain-core==0.3.34",
    "langchain-huggingface==0.1.2",
    "langchain-ollama==0.2.3",
    "langchain-text-splitters==0.3.6",
    "langsmith==0.3.8",
    "markdown-it-py==3.0.0",
    "MarkupSafe==3.0.2",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "mmh3==5.1.0",
    "monotonic==1.6",
    "mpmath==1.3.0",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "networkx==3.4.2",
    "numpy==1.26.4",
    "oauthlib==3.2.2",
    "ollama==0.4.7",
    "onnxruntime==1.20.1",
    "opentelemetry-api==1.30.0",
    "opentelemetry-exporter-otlp-proto-common==1.30.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.30.0",
    "opentelemetry-instrumentation==0.51b0",
    "opentelemetry-instrumentation-asgi==0.51b0",
    "opentelemetry-instrumentation-fastapi==0.51b0",
    "opentelemetry-proto==1.30.0",
    "opentelemetry-sdk==1.30.0",
    "opentelemetry-semantic-conventions==0.51b0",
    "opentelemetry-util-http==0.51b0",
    "orjson==3.10.15",
    "overrides==7.7.0",
    "packaging==24.2",
    "pillow==11.1.0",
    "pluggy==1.5.0",
    "posthog==3.11.0",
    "propcache==0.2.1",
    "protobuf==5.29.3",
    "pyasn1==0.6.1",
    "pyasn1_modules==0.4.1",
    "pydantic==2.10.6",
    "pydantic-settings==2.7.1",
    "pydantic_core==2.27.2",
    "Pygments==2.19.1",
    "PyPika==0.48.9",
    "pytest==8.3.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "PyYAML==6.0.2",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rich==13.9.4",
    "rpds-py==0.22.3",
    "rsa==4.9",
    "safetensors==0.5.2",
    "scikit-learn==1.6.1",
    "scipy==1.15.1",
    "sentence-transformers==3.4.1",
    "setuptools==75.8.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "SQLAlchemy==2.0.38",
    "starlette==0.45.3",
    "sympy==1.13.1",
    "tenacity==9.0.0",
    "threadpoolctl==3.5.0",
    "tokenizers==0.21.0",
    "torch==2.6.0",
    "tqdm==4.67.1",
    "transformers==4.48.3",
    "typer==0.15.1",
    "typing-inspect==0.9.0",
    "typing_extensions==4.12.2",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "uvloop==0.21.0",
    "watchdog==6.0.0",
    "watchfiles==1.0.4",
    "websocket-client==1.8.0",
    "websockets==14.2",
    "wrapt==1.17.2",
    "yarl==1.18.3",
    "zipp==3.21.0",
    "zstandard==0.23.0",
]
classifiers = [
    "Programming Language:: Python:: 3",
    "License:: OSI Approved:: GPLv3",
    "Operating System:: OS Independent",
    "AILLM:: Local",
]
readme = "README.md"
keywords = ["ai", "llm", "rag"]
license = {text = "GPLv3"}

[tool.setuptools]
include-package-data = true